REFACTOR FRAGMENT BY SUPERCLASS SUPERFRAGMENT
