+++ app/src/androidTest/java/ch/epfl/sweng/zuluzulu/TestWithAdminLogin.java
@@ -2,10 +2,20 @@ package ch.epfl.sweng.zuluzulu;
 
 import org.junit.Before;
 
+import ch.epfl.sweng.zuluzulu.Structure.User;
+
 public class TestWithAdminLogin extends TestWithLogin {
+
+    private User admin;
+
     @Before
-    public void setUp() {
-        this.user = Utility.createTestAdmin();
-        Utility.addUserToMainIntent(mActivityRule, this.user);
+    public void setUpLogin() {
+        this.admin = Utility.createTestAdmin();
+        Utility.addUserToMainIntent(mActivityRule, this.admin);
+    }
+
+    @Override
+    protected User getUser() {
+        return admin;
     }
 }
