+++ app/src/main/java/ch/epfl/sweng/zuluzulu/Fragments/SuperFragment.java
@@ -1,0 +1,28 @@
+package ch.epfl.sweng.zuluzulu.Fragments;
+
+import android.content.Context;
+import android.support.v4.app.Fragment;
+
+import ch.epfl.sweng.zuluzulu.OnFragmentInteractionListener;
+
+public class SuperFragment extends Fragment {
+    protected OnFragmentInteractionListener mListener;
+
+
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        if (context instanceof OnFragmentInteractionListener) {
+            mListener = (OnFragmentInteractionListener) context;
+        } else {
+            throw new RuntimeException(context.toString()
+                    + " must implement OnFragmentInteractionListener");
+        }
+    }
+
+    @Override
+    public void onDetach() {
+        super.onDetach();
+        mListener = null;
+    }
+}
